Perform comparison on two datasets

Perform comparison on two datasets

Usage

proc_compare_df(data_x, data_y, by, by_col = NULL, cols = NULL)

Arguments

data_x

'current' or 'new' dataset

data_y

'previous' or 'old' dataset

by

supplied join columns

by_col

unique identifier

cols

columns to compare

Value

compdf_list list of tables

Examples

# using dfdiffs::diff_modified_data
diff_modified_data <- dfdiffs::diff_modified_data
diff_current_modified <- diff_modified_data$diff_current_modified
diff_previous_modified <- diff_modified_data$diff_previous_modified
proc_compare_df(data_x = diff_current_modified,
                data_y = diff_previous_modified,
                by = c("subject_id", "record"),
                cols = c("text_value_a", "text_value_b"))
#> 
#> 
#> Table: Summary of data.frames
#> 
#> version   arg            ncol   nrow
#> --------  ------------  -----  -----
#> x         join_data_x       5      5
#> y         join_data_y       5      5
#> 
#> 
#> 
#> Table: Summary of overall comparison
#> 
#> statistic                                                      value
#> ------------------------------------------------------------  ------
#> Number of by-variables                                             1
#> Number of non-by variables in common                               4
#> Number of variables compared                                       4
#> Number of variables in x but not y                                 0
#> Number of variables in y but not x                                 0
#> Number of variables compared with some values unequal              2
#> Number of variables compared with all values equal                 2
#> Number of observations in common                                   5
#> Number of observations in x but not y                              0
#> Number of observations in y but not x                              0
#> Number of observations with some compared variables unequal        3
#> Number of observations with all compared variables equal           2
#> Number of values unequal                                           3
#> 
#> 
#> 
#> Table: Variables not shared
#> 
#>                          
#>  ------------------------
#>  No variables not shared 
#>  ------------------------
#> 
#> 
#> 
#> Table: Other variables not compared
#> 
#>                                  
#>  --------------------------------
#>  No other variables not compared 
#>  --------------------------------
#> 
#> 
#> 
#> Table: Observations not shared
#> 
#>                             
#>  ---------------------------
#>  No observations not shared 
#>  ---------------------------
#> 
#> 
#> 
#> Table: Differences detected by variable
#> 
#> var.x          var.y            n   NAs
#> -------------  -------------  ---  ----
#> subject_id     subject_id       0     0
#> record         record           0     0
#> text_value_a   text_value_a     2     0
#> text_value_b   text_value_b     1     0
#> 
#> 
#> 
#> Table: Differences detected
#> 
#> var.x          var.y          JOINCOL   values.x                             values.y            row.x   row.y
#> -------------  -------------  --------  -----------------------------------  -----------------  ------  ------
#> text_value_a   text_value_a   A-1       Issue resolved                       Issue unresolved        1       1
#> text_value_a   text_value_a   A-2       Issue resolved                       Issue unresolved        2       2
#> text_value_b   text_value_b   C-4       Joint pain, stiffness and swelling   Joint pain              4       4
#> 
#> 
#> 
#> Table: Non-identical attributes
#> 
#>                              
#>  ----------------------------
#>  No non-identical attributes 
#>  ----------------------------
#>